pyproject.toml
[project]
name = "vantage-mcp"
version = "0.1.0"
description = "Black-box web pentesting MCP server"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=3.0.2",
    "httpx>=0.28",
    "beautifulsoup4>=4.13",
    "lxml>=5.3",
    "openai>=1.60",
    "python-dotenv>=1.0",
]
